refactor: auth-controller package
