+++ waltid-services/waltid-wallet-api/src/main/kotlin/id/walt/webwallet/service/account/EmailAccountStrategy.kt
@@ -4,7 +4,7 @@ import de.mkammerer.argon2.Argon2Factory
 import id.walt.commons.web.ConflictException
 import id.walt.commons.web.UnauthorizedException
 import id.walt.webwallet.db.models.Accounts
-import id.walt.webwallet.web.controllers.ByteLoginRequest
+import id.walt.webwallet.web.controllers.auth.ByteLoginRequest
 import id.walt.webwallet.web.model.EmailAccountRequest
 import kotlinx.datetime.Clock
 import kotlinx.datetime.toJavaInstant
