+++ waltid-services/waltid-wallet-api/src/main/kotlin/id/walt/webwallet/web/controllers/IssuersController.kt
@@ -3,6 +3,7 @@ package id.walt.webwallet.web.controllers
 import id.walt.webwallet.service.WalletServiceManager
 import id.walt.webwallet.service.issuers.CredentialDataTransferObject
 import id.walt.webwallet.service.issuers.IssuerDataTransferObject
+import id.walt.webwallet.web.controllers.auth.getWalletService
 import io.github.smiley4.ktorswaggerui.dsl.routing.get
 import io.github.smiley4.ktorswaggerui.dsl.routing.post
 import io.github.smiley4.ktorswaggerui.dsl.routing.put
