+++ waltid-services/waltid-wallet-api/src/main/kotlin/id/walt/webwallet/web/controllers/ReportsController.kt
@@ -2,6 +2,8 @@ package id.walt.webwallet.web.controllers
 
 import id.walt.webwallet.db.models.WalletCredential
 import id.walt.webwallet.service.report.CredentialReportRequestParameter
+import id.walt.webwallet.web.controllers.auth.getWalletId
+import id.walt.webwallet.web.controllers.auth.getWalletService
 import io.github.smiley4.ktorswaggerui.dsl.routing.get
 import io.github.smiley4.ktorswaggerui.dsl.routing.route
 import io.ktor.http.*
